+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+<!-- markdownlint-configure-file { "MD024": { "siblings_only": true } } -->
+
 # Change Log
 
 All notable changes to this project will be documented in this file.
@@ -11,10 +13,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- *hotspot*のスタイルを上書きできるように変更
+- *panel*のスタイルを上書きできるように変更
+- カメラを初期位置に戻すボタンのスタイルを上書きできるように変更
+- スクリーンショットボタンのスタイルを上書きできるように変更
+
 ### Deprecated
 
 ### Removed
 
 ### Fixed
 
-[unreleased]: https://github.com/cynack/figni-viewer/compare/v1.0.0...HEAD
+## [0.0.13] - 2021-11-18
+
+### Added
+
+- *hotspot*の表示非表示を変化させる`to-state`属性を追加
+- *hotspot*の表示非表示を決定する`visible`属性を追加